Quickfix to prevent transaction performance collector OOM
